refactor: split class
